feat(api): preserve NIfTI dtype, lock tiling arithmetic, rename extract_zarr_to_image
- convert_nifti_to_zarr: replace get_fdata() with np.asanyarray(ni_img.dataobj)
  to preserve the stored dtype (uint16 stays uint16, no float64 upcast / 4x
  storage inflation)
- predict_one: wrap n_patches_by_row in int() so y1 % n_patches_by_row is
  integer arithmetic with no float rounding (locks tiling path for the
  future patching=True implementation)
- extract_zarr_to_png -> extract_zarr_to_image: clean-break rename + format
  param (default 'tiff' writes a single multi-page TIFF via tifffile.imwrite;
  'png' escape hatch writes per-slice PNGs; unsupported format raises
  ValueError)
- utils/__init__.py: update barrel import + __all__ to extract_zarr_to_image
- add tifffile as a direct core dependency (previously only transitive via
  scikit-image)
- tests: dtype-preservation regression (conversion), int-modulo tiling
  arithmetic (prediction, torch-gated), TIFF round-trip + PNG escape hatch +
  unsupported-format ValueError (io)

+++ liom_toolkit/segmentation/vseg/prediction.py
@@ -146,7 +146,12 @@ def predict_one(
     """ Load dataset """
     test_x = numeric_filesort(f"{save_path}/patches", folder="images")
 
-    n_patches_by_row = (processed_image.shape[1] - W) / stride + 1
+    # Wrap in int(...) so the ``y1 % n_patches_by_row`` modulo below is
+    # integer arithmetic with no float rounding. This locks the tiling-path
+    # arithmetic for the future patching=True implementation (currently
+    # raises NotImplementedError); a float modulo would silently round wrong
+    # on non-evenly-divisible dimensions.
+    n_patches_by_row = int((processed_image.shape[1] - W) / stride + 1)
 
     x1 = 0
     y1 = 0
